+++ certlogic/certlogic-js/src/validation/data-accesses.ts
@@ -1,8 +1,13 @@
 import { CertLogicExpression, CertLogicOperation } from "../typings"
 
 
+type DataAccessWithContext = {
+    path: string
+    context: CertLogicExpression
+}
+
 const gatherFrom = (expr: CertLogicExpression, parent?: CertLogicExpression): DataAccessWithContext[] => {
-    const recurse = (subExpr: CertLogicExpression) => gatherFrom(subExpr, expr || subExpr)
+    const recurse = (subExpr: CertLogicExpression) => gatherFrom(subExpr, expr)
     if (Array.isArray(expr)) {
         return (expr as CertLogicExpression[]).flatMap(recurse)
     }
@@ -13,7 +18,7 @@ const gatherFrom = (expr: CertLogicExpression, parent?: CertLogicExpression): Da
         switch (operator) {
 
             case "var": return [ { path: values, context: parent ?? expr } ]
-            case "if": return values.slice(0, 3).flatMap(recurse)    // guard, then, else
+            case "if": return values.slice(0, 3).flatMap(recurse)    // 0: guard, 1: then, 2: else
 
             // all infixes:
             case "===":
@@ -32,7 +37,11 @@ const gatherFrom = (expr: CertLogicExpression, parent?: CertLogicExpression): Da
             case "!": return recurse(values[0])
             case "plusTime": return recurse(values[0])
             case "reduce": return [ /* operand: */values[0], /* initial: */values[2] ].flatMap(recurse)
-            case "extractFromUVCI": return [ /* operand: */values[0], /* index: */values[1] ].flatMap(recurse)
+            case "extractFromUVCI": return recurse(values[0])
+
+            /*
+             * Note: Array.slice and recurse(Array[index]) are resilient against missing values.
+             */
 
             default:
                 throw new Error(`operator not recognised by fields gatherer ("gatherFrom") in certlogic-js/validation/${__filename}: "${operator}"`)
@@ -48,11 +57,6 @@ const gatherFrom = (expr: CertLogicExpression, parent?: CertLogicExpression): Da
  */
 export const dataAccesses = (expr: CertLogicExpression) => [ ...new Set(gatherFrom(expr).map((dataAccess) => dataAccess.path)) ].sort()
 
-
-type DataAccessWithContext = {
-    path: string
-    context: CertLogicExpression
-}
 
 /**
  * A map with paths accessed as keys,
@@ -76,8 +80,8 @@ export type DataAccessesWithContext = { [path: string]: CertLogicExpression[] }
 /**
  * Compute which data accesses can be performed by the given CertLogic expression,
  * including in which contexts that would then happen.
- * The context consists of the CertLogic expression with an operand that performs the data access (= "`var`" operation),
- * or that "`var`" operation when its the entire expression.
+ * The context consists of the CertLogic expression with an operand that performs the data access (= `var` operation),
+ * or that `var` operation when it's the entire expression.
  * @param expr A CertLogic expression.
  */
 export const dataAccessesWithContext = (expr: CertLogicExpression): DataAccessesWithContext => {