+++ certlogic/certlogic-js/CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.10.0
 
 * Add a `dataAccessesWithContext` function in the `validation` sub package
+* Provide better format validations error in case an operation has no values/operands
 
 
 ## 0.9.2