+++ certlogic/certlogic-kotlin/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+# 0.10.0
+
+* Add a `dataAccesses` Kotlin file with `dataAccesses` and `dataAccessesWithContext` functions
+* Provide better format validations error in case an operation has no values/operands
+
 
 ## 0.9.0